requirements.txt
Django>=4.2,<5.0
python-dotenv>=1.0.0
djangorestframework>=3.14
django-cors-headers>=4.0
django-filter>=23.0
Pillow>=10.0
qrcode[pil]>=7.4
mysqlclient>=2.2.0
PyJWT>=2.8.0
